t: Adopted blocks={-1,-1} for clobbered/invalid mdirs
blocks={0,0} technically worked, but only because the only mdirs allowed
at block 0 are mroot blocks. In theory, an mdir at block 0 could match
blocks={0,0} and clobber incorrectly, but it's not possible for such an
mdir to be in a non-trivial mtree, since blocks={0,1} are reserved for
the mrootanchor.

But bleh, that's complicated.

Setting blocks={-1,-1} makes the mdir truely invalid/unmatchable and
provides a stronger invariant at a minor code cost.

Code changes:

           code          stack
  before: 34554           2624
  after:  34566 (+0.0%)   2624 (+0.0%)