Switched to using did for dir-is-empty checks, lfsr_grm_pushdid
So instead of looking for another bookmark, we check to see if the next
mid's did matches.

This is in theory more robust, and can handle neighboring
directories/files without bookmarks, but adds a bit of code cost.

Deduplicating these checks into lfsr_grm_pushdid helps reduce this a
bit:

                   code          stack
  before:         34406           2624
  after-no-dedup: 34526 (+0.3%)   2624 (+0.0%)
  after-dedup:    34470 (+0.2%)   2624 (+0.0%)