t: Clobber traversals at the omdir level, rely on state machine
This brings back clobbering individual omdirs, so modifying an unsynced
file should leave all other traversals intact. This is the most precise
level of clobbering that I think is reasonable to implement.

This means we should be able to, say, check all currently-committed
checksums while writing to unsynced files at the same time. Which might
be useful? Maybe?

To make this work, lfsr_traversal_clobber now relies on the current
traversal state to know how to clobber correctly. This is more verbose,
but likely safer/more flexible.

Curiously, this actually ended up saving a bit of code, which is a bit
surprising:

           code          stack
  before: 35356           2688
  after:  35304 (-0.1%)   2688 (+0.0%)

Maybe manipulating the state machine directly gives the compiler more
info to work with? Not sure.