t: Added lfsr_fs_mkdirty to mark all traversals as dirty
Does what it says on the tin.

This simplifies lfsr_omdir_clobber, and can be used in more places. It's
a bit more flexible than implicitly mkdirtying in lfs_alloc_ckpoint, but
does add another function call.

But thanks to better code reuse this ends up saving a bit of code:

           code          stack
  before: 35304           2688
  after:  35264 (-0.1%)   2680 (-0.3%)